+++ tests/test_element_handle.py
@@ -328,6 +328,42 @@ class TestQuerySelector(BaseTestCase):
         self.assertEqual(len(elements), 0)
 
     @sync
+    async def test_element_handle_JJEval(self):
+        await self.page.setContent(
+            '<html><body><div class="tweet"><div class="like">100</div>'
+            '<div class="like">10</div></div></body></html>'
+        )
+        tweet = await self.page.J('.tweet')
+        content = await tweet.JJeval(
+            '.like', 'nodes => nodes.map(n => n.innerText)')
+        self.assertEqual(content, ['100', '10'])
+
+    @sync
+    async def test_element_handle_JJEval_subtree(self):
+        await self.page.setContent(
+            '<div class="a">not-a-child-div</div>'
+            '<div id="myId">'
+            '<div class="a">a1-child-div</div>'
+            '<div class="a">a2-child-div</div>'
+            '</div>'
+        )
+        elementHandle = await self.page.J('#myId')
+        content = await elementHandle.JJeval(
+            '.a', 'nodes => nodes.map(n => n.innerText)')
+        self.assertEqual(content, ['a1-child-div', 'a2-child-div'])
+
+    @sync
+    async def test_element_handle_JJEval_missing_selector(self):
+        await self.page.setContent(
+            '<div class="a">not-a-child-div</div><div id="myId"></div>')
+        elementHandle = await self.page.J('#myId')
+        with self.assertRaises(ElementHandleError) as cm:
+            await elementHandle.JJeval(
+                '.a', 'nodes => nodes.map(n => n.innerText)')
+        self.assertIn('Error: failed to find elements matching selector ".a"',
+                      cm.exception.args[0])
+
+    @sync
     async def test_element_handle_xpath(self):
         await self.page.setContent(
             '<html><body><div class="second"><div class="inner">A</div></div></body></html>'  # noqa: E501